spider/requirements.txt
Scrapy==2.7.1
scrapy-splash==0.8.0
pyzmq==25.0.0